[PT2E][Quant][BE] Split short term and long term tests in different files
Just for better organization

Differential Revision: [D44918492](https://our.internmc.facebook.com/intern/diff/D44918492/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D44918492/)!

[ghstack-poisoned]
